+++ packages/compass-components/src/index.ts
@@ -19,6 +19,10 @@ import FileInput from './components/file-input';
 import { Modal } from './components/modal';
 import { ModalTitle } from './components/modal-title';
 import { Toolbar } from './components/toolbar';
+import {
+  ErrorSummary,
+  WarningSummary,
+} from './components/error-warning-summary';
 import { RadioBoxGroup } from './components/radio-box-group';
 import { SpinLoader } from './components/spin-loader';
 import { ResizeHandle, ResizeDirection } from './components/resize-handle';
@@ -54,6 +58,7 @@ export {
   Accordion,
   CancelLoader,
   ConfirmationModal,
+  ErrorSummary,
   FileInput,
   Modal,
   ModalTitle,
@@ -62,6 +67,7 @@ export {
   ResizeHandle,
   ResizeDirection,
   Toolbar,
+  WarningSummary,
   WorkspaceTabs,
 };
 export {
