+++ packages/compass-home/src/index.less
@@ -21,17 +21,6 @@
   }
 }
 
-.rtss-databases,
-.rt-perf {
-  //rtss charts
-  display: flex;
-  flex-direction: column;
-  align-items: stretch;
-  width: 100%;
-  flex-grow: 1;
-  overflow: auto;
-}
-
 .controls-container {
   flex-grow: 0;
   flex-shrink: 0;
