+++ packages/compass-serverstats/package.json
@@ -30,7 +30,6 @@
     "@mongodb-js/compass-components": "^0.13.0",
     "d3": "^3.5.17",
     "hadron-react-buttons": "^5.7.0",
-    "hadron-react-components": "^5.13.0",
     "react": "^16.14.0",
     "react-dom": "^16.14.0"
   },
@@ -61,7 +60,6 @@
     "hadron-app": "^4.21.0",
     "hadron-app-registry": "^8.10.0",
     "hadron-react-buttons": "^5.7.0",
-    "hadron-react-components": "^5.13.0",
     "mocha": "^8.4.0",
     "mongodb-connection-model": "^21.15.0",
     "mongodb-data-service": "^21.19.0",
