+++ packages/compass-serverstats/src/components/index.less
@@ -2,9 +2,7 @@
 
 @import './rtss-db.less';
 @import './rtss-lists.less';
-@import './rtss-errors.less';
 @import './rtss-charts.less';
-@import './rtss-databases.less';
 
 .rtss {
   font-family: "Akzidenz", "Helvetica Neue", Helvetica, Arial, sans-serif;
