+++ packages/compass-serverstats/src/components/rtss-charts.less
@@ -1,4 +1,12 @@
 .rt-perf {
+  display: flex;
+  flex-direction: column;
+  align-items: stretch;
+  width: 100%;
+  flex-grow: 1;
+  overflow: auto;
+  background: #3D4247;
+
   .serverstats {
     display: flex;
     flex-direction: column;
