--- packages/compass-serverstats/src/components/rtss-databases.less
@@ -1,31 +1,0 @@
-.rtss-databases {
-  background-color: @gray8;
-
-  &-link, &-link:hover, &-link:active {
-    font-weight: bold;
-  }
-
-  &-create-button {
-    height: 32px;
-    background: @gray8;
-    border-bottom: 1px solid @gray7;
-    padding-top: 4px;
-    padding-left: 15px;
-  }
-}
-
-.no-collections-zero-state {
-  font-size: 18px;
-  color: @gray4;
-  font-weight: 200;
-  position: absolute;
-  left: ~"calc(50% - 250px)";
-  right: 15%;
-  top: 50%;
-  z-index: 2;
-
-  a {
-    cursor: pointer;
-    margin: 0.5ex;
-  }
-}
