--- packages/compass-serverstats/src/components/rtss-errors.less
@@ -1,16 +1,0 @@
-.rt-errors {
-}
-.rt-errors__list {
-  padding-left: 0;
-  margin-bottom: 0;
-}
-.rt-errors__item {
-  position: relative;
-  overflow: hidden;
-}
-.rt-errors__item:last-child {
-  border-bottom: none;
-}
-.rt-errors__bold {
-  font-weight: 900;
-}
