+++ packages/connection-form/src/components/connect-form-actions.tsx
@@ -2,11 +2,12 @@ import React from 'react';
 import {
   Button,
   ButtonVariant,
+  ErrorSummary,
+  WarningSummary,
   spacing,
   uiColors,
   css,
 } from '@mongodb-js/compass-components';
-import { ErrorSummary, WarningSummary } from './validation-summary';
 import type {
   ConnectionFormError,
   ConnectionFormWarning,
@@ -41,8 +42,10 @@ function ConnectFormActions({
 }): React.ReactElement {
   return (
     <div className={formActionStyles}>
-      {warnings.length ? <WarningSummary warnings={warnings} /> : ''}
-      {errors.length ? <ErrorSummary errors={errors} /> : ''}
+      {warnings && <WarningSummary warnings={warnings} />}
+      {errors && (
+        <ErrorSummary dataTestId="connection-error-summary" errors={errors} />
+      )}
       <div className={formActionButtonsStyles}>
         {saveButton !== 'hidden' && (
           <Button
